base-requirements.txt
Faker==1.0.7
Flask>=1.1.2,<=2.0.0
Flask-Caching==1.10.0,<=2.0.0
Flask-Compress==1.0.2
Flask-Cors==3.0.9
Flask-Script==2.0.5
flask-talisman==0.7.0
manage.py==0.2.10
pylibmc==1.5.2
meza>=0.42.5,<=1.0.0
mezmorize>=0.28.0,<=1.0.0
riko>=0.62.0,<=1.0.0
pygogo>=0.13.0,<=2.0.0
requests-oauthlib==1.2.0
oauthlib==3.0.1
inflect==4.1.0
python-dotenv>=0.14.0,<=1.0.0
attrs>=21.4.0,<=22.0.0
playwright>=1.30.0,<=2.0.0
